hydra/lib/pyproject.toml
[tool.poetry]
name = "laniakea"
version = "0.2.2"
description = "Laniakea Python Library"
authors = ["Josselin Pujo <pujo.j@sfeir.com>"]
license = "Apache-2.0"

[tool.poetry.scripts]
lnk = 'laniakea.cli:main'

[tool.poetry.dependencies]
python = "~3.8.2"

# IO
requests = { version ="^2.22.0" }
lxml = { version = "^4.5.2" }
beautifulsoup4 = { version = "^4.9.1" }
pysftp = { version = "^0.2.9" }
cx-Oracle = { version = "^8.0.0", markers = "sys_platform == 'linux'" }
ctds = { version = "^1.12.0", markers = "sys_platform == 'linux'" }
psycopg2 = { version = "^2.8.5", markers = "sys_platform == 'linux'" }
mysqlclient = { version = "^2.0.1", markers = "sys_platform == 'linux'" }
SQLAlchemy = { version = "^1.3.19" }
kafka-python = { version = "^2.0.1" }
azure-servicebus = { version = "^0.50.3" }
azure-cosmos = { version = "^4.1.0" }
gcsfs = { version = "^0.7.0" }

# Google Cloud
google-cloud-storage = { version = "^1.24.0" }
google-cloud-bigquery-storage = { extras = ["pandas", "fastavro", "pyarrow"], version = "^1.0.0" }
google-cloud-pubsub = { version = "^1.7.0" }
google-cloud-datacatalog = { version = "^1.0.0" }

# Data
dbt = { version = "^0.17.2" }
tensorflow = { version = "^2.3.0" }
pandas = { version = "^1.1.1" }
xgboost = { version = "^1.2.0" }
numba = { version = "^0.51.0" }
vaex = { version = "^3.0.0" }
scikit-learn = { version = "^0.23.2" }

# Export and reporting
pyppeteer = { version = "^0.2.2" }
sendgrid = { version = "^6.4.6" }
XlsxWriter = { version = "^1.3.3" }
bokeh = { version = "^2.2.0" }
dash = { version = "^1.14.0" }

# Execution env
prefect = { version = "0.13.1" }
dask = { version = "^2.24.0" }
distributed = { version = "^2.24.0" }
semver = { version = "^2.10.2" }
gitpython = { version = "^3.1.7" }
jupyterlab = { version = "^2.2.6" }
kubernetes = { version = "^11.0.0" }

[build-system]
requires = ["poetry>=1.0.10"]
build-backend = "poetry.masonry.api"
